# study-service/src/main/resources/openapi/study-api.openapi.yml
openapi: 3.0.0
info:
  title: Study API
  version: 0.0.1
servers:
  - url: http://localhost:8083/api/study
paths:
...
  /group/{id}:
    get:
      security:
        - bearerAuth: [ ]
      summary: Group Information
      tags:
        - Group
      operationId: getGroupById
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyGroupOpenAPI"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
    delete:
      security:
        - bearerAuth: [ ]
      summary: Delete a Group By Id
      tags:
        - Group
      operationId: deleteGroupById
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: Deleted
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
  /group:
    get:
      security:
        - bearerAuth: [ ]
      summary: All Groups Information
      tags:
        - Group
      operationId: getAllGroups
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudyGroupOpenAPI"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
    post:
      security:
        - bearerAuth: [ ]
      summary: Create a Group
      tags:
        - Group
      operationId: createGroup
      requestBody:
        description: Group Object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudyGroupOpenAPI'
      responses:
        '201':
          description: Created
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
  /group/user:
    post:
      security:
        - bearerAuth: [ ]
      summary: Add user in group
      tags:
        - Group
      operationId: addUserInGroup
      requestBody:
        description: User id, group id
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInGroupRequestOpenApi'
      responses:
        '201':
          description: Added
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation

    delete:
      security:
        - bearerAuth: [ ]
      summary: Delete user from group
      tags:
        - Group
      operationId: deleteUserFromGroup
      requestBody:
        description: User id, group id
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInGroupRequestOpenApi'
      responses:
        '200':
          description: Deleted
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
...
  /attendance-statistics/subject/{subject_id}:
    get:
      security:
        - bearerAuth: [ ]
      summary: Subject's Attendance Statistics Information
      tags:
        - AttendanceStatistics
      operationId: getStatisticsBySubjectId
      parameters:
        - in: path
          name: subject_id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceStatisticsOpenApi"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
...
  /subject:
    get:
      security:
        - bearerAuth: [ ]
      summary: All Subjects Information
      tags:
        - Subject
      operationId: getAllSubjects
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubjectOpenAPI"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
    post:
      security:
        - bearerAuth: [ ]
      summary: Create a Subject
      tags:
        - Subject
      operationId: createSubject
      requestBody:
        description: Subject Object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubjectOpenAPI'
      responses:
        '201':
          description: Created
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
    put:
      security:
        - bearerAuth: [ ]
      summary: Update a Subject
      tags:
        - Subject
      operationId: updateSubject
      requestBody:
        description: Subject Object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubjectOpenAPI'
      responses:
        '200':
          description: Updated
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
  /subject/{id}:
    get:
      security:
        - bearerAuth: [ ]
      summary: Subject Information
      tags:
        - Subject
      operationId: getSubjectById
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubjectOpenAPI"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found

    delete:
      security:
        - bearerAuth: [ ]
      summary: Delete a Subject by Id
      tags:
        - Subject
      operationId: deleteSubject
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: Deleted
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
  /subject/group/{group_id}:
    get:
      security:
        - bearerAuth: [ ]
      summary: Subjects Information By Group Id
      tags:
        - Subject
      operationId: getSubjectsByGroupId
      parameters:
        - in: path
          name: group_id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubjectOpenAPI"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
  /subject/creator/{creator_id}:
    get:
      security:
        - bearerAuth: [ ]
      summary: Subjects Information By Creator Id
      tags:
        - Subject
      operationId: getSubjectsByCreatorId
      parameters:
        - in: path
          name: creator_id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubjectOpenAPI"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
  /subject/editor/{editor_id}:
    get:
      security:
        - bearerAuth: [ ]
      summary: Subjects Information By Editor Id
      tags:
        - Subject
      operationId: getSubjectsByEditor
      parameters:
        - in: path
          name: editor_id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubjectOpenAPI"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
  /materials/subject/{subject_id}:
    get:
      security:
        - bearerAuth: [ ]
      summary: Get All Materials by Subject Id
      tags:
        - Materials
      operationId: getMaterialsBySubjectId
      parameters:
        - in: path
          name: subject_id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudyMaterialsOpenAPI"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
...
components:
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    StudyGroupOpenAPI:
      type: object
      required:
        - name
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        studentsIds:
          type: array
          description: studentsIDs
          items:
            type: integer
            format: int64
    SubjectOpenAPI:
      type: object
      required:
        - name
        - studyGroupsIds
        - creatorId
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        studyGroupsIds:
          type: array
          items:
            type: integer
            format: int64
        studyMaterialsIds:
          type: array
          items:
            type: integer
            format: int64
        scheduleIds:
          type: array
          items:
            type: integer
            format: int64
        creatorId:
          type: integer
          format: int64
        editorsIds:
          type: array
          items:
            type: integer
            format: int64
...
    StudyMaterialsType:
      type: string
      enum:
        - MATERIAL
        - TASK
    StudyMaterialsOpenAPI:
      type: object
      required:
        - type
        - subjectId
        - materialsText
      properties:
        id:
          type: integer
          format: int64
        type:
          $ref: "#/components/schemas/StudyMaterialsType"
        materialsText:
          type: string
        subjectId:
          type: integer
          format: int64
        homeworkIds:
          type: array
          items:
            type: integer
            format: int64
        filesId:
          type: array
          items:
            type: integer
            format: int64
        filesNames:
          type: array
          items:
            type: string
...
    UserInGroupRequestOpenApi:
      type: object
      required:
        - userId
        - groupId
      properties:
        userId:
          type: integer
          format: int64
        groupId:
          type: integer
          format: int64
...
    AttendanceStatisticsOpenApi:
      type: object
      required:
        - scheduleId
        - userId
        - status
      properties:
        scheduleId:
           type: integer
           format: int64
        status:
          $ref: "#/components/schemas/AttendanceStatusOpenApi"
        studentId:
          type: integer
          format: int64
    AttendanceStatusOpenApi:
      type: string
      enum:
        - ATTENDED
        - MISSED
...